Untrustless MPC for Prod BSC

+++ relay-mpc.sh
@@ -1,30 +1,49 @@
 #!/bin/bash
 set -e
 
-apt-get remove -y docker docker-engine docker.io containerd runc && \
-apt-get update && \
-apt-get install -y \
-    ca-certificates \
-    curl \
-    gnupg \
-    lsb-release
+# TODO: remove this after update users' relays; needed for moving to new version of the script
+[ -d ${PWD}/share ] && mv share share_eth
 
-mkdir -p /etc/apt/keyrings
 
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+SIDE_NET=$1
+MPC_MEID=$2
+MPC_ACCESS_TOKEN=$3
+MPC_PARTY_IDS=$4
+MPC_THRESHOLD=$5
+MPC_KEYGEN_URL=$6
 
-echo \
-  "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
-  $(lsb_release -cs) stable" | tee /etc/apt/sources.list.d/docker.list > /dev/null
+SIDE_NET_LOWERCASE="${SIDE_NET,,}"
+SIDE_NET_UPPERCASE="${SIDE_NET^^}"
 
-apt-get update
-apt-get install -y docker-ce docker-ce-cli containerd.io docker-compose-plugin
+RELAY_CONTAINER_NAME="${SIDE_NET_LOWERCASE}-relay"
+KEYGEN_CONTAINER_NAME="${SIDE_NET_LOWERCASE}-relay-keygen"
 
-MPC_MEID=$1
-MPC_ACCESS_TOKEN=$2
-MPC_PARTY_IDS=$3
-MPC_THRESHOLD=$4
-MPC_KEYGEN_URL=$5
+SHARE_DIR="${PWD}/share_${SIDE_NET_LOWERCASE}"
+SHARE_PATH="${SHARE_DIR}/share_${MPC_MEID}"
+
+
+# install docker by default
+if [[ -z $NO_DOCKER_INSTALL ]]; then
+    apt-get remove -y docker docker-engine docker.io containerd runc && \
+    apt-get update && \
+    apt-get install -y \
+        ca-certificates \
+        curl \
+        gnupg \
+        lsb-release
+
+    mkdir -p /etc/apt/keyrings
+
+    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+
+    echo \
+      "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
+      $(lsb_release -cs) stable" | tee /etc/apt/sources.list.d/docker.list > /dev/null
+
+    apt-get update
+    apt-get install -y docker-ce docker-ce-cli containerd.io docker-compose-plugin
+fi
+
 
 echo "Please enter your private key"
 
@@ -38,12 +57,12 @@ do
     fi
 done
 
-read -sp 'Ethereum private key: ' ETH_PRIVATE_KEY
+read -sp "${SIDE_NET} network private key: " SIDE_PRIVATE_KEY
 while true;
 do
     echo -e "\n"
-    if [ ${#ETH_PRIVATE_KEY} -ne 64 ];
-        then read -sp 'Key length should be 64 characters, type again: ' ETH_PRIVATE_KEY;
+    if [ ${#SIDE_PRIVATE_KEY} -ne 64 ];
+        then read -sp 'Key length should be 64 characters, type again: ' SIDE_PRIVATE_KEY;
         else break
     fi
 done
@@ -64,7 +83,7 @@ done
 
 
 set +e
-docker rm -f eth-relay
+docker rm -f $RELAY_CONTAINER_NAME
 set -e
 
 IMAGE=ghcr.io/ambrosus/ambrosus-bridge
@@ -80,12 +99,12 @@ fi
 docker pull $IMAGE:$TAG
 
 echo "Starting keygen..."
-if [ -f "$PWD/share/share_$MPC_MEID" ]; then
+if [ -f $SHARE_PATH ]; then
   echo "Share already exist"
 else
   docker run -it --rm \
-  --name eth-relay-keygen \
-  -v $PWD/share:/app/shared \
+  --name $KEYGEN_CONTAINER_NAME \
+  -v $SHARE_DIR:/app/shared \
   --entrypoint '/bin/sh' \
   $IMAGE:$TAG \
   -c "go run ./cmd/mpc_keygen -url $MPC_KEYGEN_URL -meID $MPC_MEID -partyIDs '$MPC_PARTY_IDS' -threshold $MPC_THRESHOLD -accessToken $MPC_ACCESS_TOKEN -shareDir ./shared"
@@ -93,13 +112,17 @@ fi
 
 echo "Starting relay..."
 docker run -d \
---name eth-relay \
+--name $RELAY_CONTAINER_NAME \
 --restart unless-stopped \
--v $PWD/share:/app/shared \
+-v $SHARE_DIR:/app/shared \
 -e STAGE=$STAGE \
--e NETWORK=eth-untrustless \
+-e NETWORK="${SIDE_NET_LOWERCASE}-untrustless" \
 -e NETWORKS_AMB_PRIVATEKEY=$AMB_PRIVATE_KEY \
--e NETWORKS_ETH_PRIVATEKEY=$ETH_PRIVATE_KEY \
+-e "NETWORKS_${SIDE_NET_UPPERCASE}_PRIVATEKEY"=$SIDE_PRIVATE_KEY \
+-e NETWORKS_AMB_HTTPURL=$AMB_HTTP_URL \
+-e "NETWORKS_${SIDE_NET_UPPERCASE}_HTTPURL"=$SIDE_HTTP_URL \
+-e NETWORKS_AMB_WSURL=$AMB_WS_URL \
+-e "NETWORKS_${SIDE_NET_UPPERCASE}_WSURL"=$SIDE_WS_URL \
 -e EXTERNALLOGGER_TELEGRAM_TOKEN=$EXTERNALLOGGER_TELEGRAM_TOKEN \
 -e SUBMITTERS_AMBTOSIDE_MPC_MEID=$MPC_MEID \
 -e SUBMITTERS_AMBTOSIDE_MPC_SHAREPATH="shared/share_$MPC_MEID" \
@@ -110,4 +133,4 @@ docker run -d \
 $IMAGE:$TAG >> /dev/null
 
 sleep 10
-docker logs eth-relay
+docker logs $RELAY_CONTAINER_NAME
